+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Initial release of the package
 ## 1.0.1
 Fixed imports issue
 
-#1.0.41
+## 1.0.41
 * Refactored package methods to reduce package size
 * Added method to query transactions
+
+# 1.0.42
+* Documentation
+* Renamed method [getDataBundleOptions]
+* More error checking